refactor: ディレクトリをクリーンアーキテクチャ風にリファクタ
