+++ server/src/domain/model/book.go
