+++ server/src/middleware/bookMiddleware.go
